Fixed some bugs to make GCC work. I needed to change SimpleValueAction to take the function as a variable as opposed to a template parameter - GCC won't accept pointers from a base class in templates.
git-svn-id: svn://svn.code.sf.net/p/magicseteditor/code/trunk@231 0fc631ac-6414-0410-93d0-97cfa31319b6

+++ src/data/font.cpp
@@ -48,7 +48,7 @@ void reflect_font(Reader& tag, Font& font) {
 	REFLECT(weight);
 	REFLECT(style);
 	if (!name.empty())   font.font.SetFaceName(name);
-	if (size > 0)        font.font.SetPointSize(font.size = size);
+	if (size > 0)        font.font.SetPointSize((int) (font.size = size));
 	if (!weight.empty()) font.font.SetWeight(weight == _("bold")   ? wxBOLD   : wxNORMAL);
 	if (!style.empty())  font.font.SetWeight(style  == _("italic") ? wxITALIC : wxNORMAL);
 }
