+++ src/render/value/image.cpp
@@ -70,7 +70,7 @@ void ImageValueViewer::onStyleChange() {
 
 void ImageValueViewer::loadMask(const Rotation& rot) const {
 	if (style().mask_filename().empty()) return; // no mask
-	int w = rot.trS(style().width), h = rot.trS(style().height);
+	int w = (int) rot.trS(style().width), h = (int) rot.trS(style().height);
 	if (alpha_mask && alpha_mask->size == wxSize(w,h)) return; // mask loaded and right size
 	// (re) load the mask
 	Image image;
